Coverage visualization geom

Visualize quantitative genomic signal as line, area, or heatmap tiles. Input data must contain genomic coordinates (start, end) and a numeric signal value (score). The geom automatically maps these columns to the required aesthetics for the chosen type.

Usage

geom_coverage(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  type = "area",
  ...,
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE
)

Arguments

mapping

Set of aesthetic mappings created by aes(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data

The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

stat

The statistical transformation to use on the data for this layer. When using a geom_*() function to construct a layer, the stat argument can be used to override the default coupling between geoms and stats. The stat argument accepts the following:

A Stat ggproto subclass, for example StatCount.

A string naming the stat. To give the stat as a string, strip the function name of the stat_ prefix. For example, to use stat_count(), give the stat as "count".

For more information and other ways to specify the stat, see the layer stat documentation.

position

A position adjustment to use on the data for this layer. This can be used in various ways, including to prevent overplotting and improving the display. The position argument accepts the following:

The result of calling a position function, such as position_jitter(). This method allows for passing extra arguments to the position.

A string naming the position adjustment. To give the position as a string, strip the function name of the position_ prefix. For example, to use position_jitter(), give the position as "jitter".

For more information and other ways to specify the position, see the layer position documentation.

type

Visualization style: "area" (default), "line", or "heatmap".

"area"/"line": score is mapped to y (height).

"heatmap": score is mapped to fill, producing colored tiles.

...

Additional arguments passed to ggplot2::layer(), e.g. linewidth = 0.8, or alpha = 0.6. Note: color and colour parameters are ignored as this geom only uses fill aesthetics.

na.rm

If TRUE, silently drop NA values.

show.legend

logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display. To include legend keys for all levels, even when no data exists, use TRUE. If NA, all levels are shown in legend, but unobserved levels are omitted.

inherit.aes

If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. annotation_borders().

Value

A ggplot2 layer that can be added to a plot.

Examples

if (FALSE) { # \dontrun{
library(ggplot2)
df <- data.frame(
  start = seq(1, 100, 10),
  end   = seq(10, 100, 10),
  score = rnorm(10)
)

# Area plot (default)
ggplot(df) +
  geom_coverage()

# Line plot
ggplot(df) +
  geom_coverage(type = "line")

# Heatmap tiles
ggplot(df) +
  geom_coverage(type = "heatmap") +
  scale_fill_viridis_c()
} # }