A theme for sashimi plots

This function creates a theme specifically for sashimi plots that combine coverage tracks with junction arcs. It provides extra vertical margin for arc labels and handles y-axis display options.

Usage

ez_sashimi_theme(y_axis_style = c("none", "simple", "full"), ...)

Arguments

y_axis_style

Style of the y-axis. Options are "none", "simple", and "full". Default: "none"

...

Additional arguments passed to ez_theme

Value

A ggplot2 theme object

Examples

if (FALSE) { # \dontrun{
library(ggplot2)
p <- ggplot(data, aes(x = start, y = score)) +
  geom_area() +
  ez_sashimi_theme()
} # }